# web/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Laplace.Endpoints.OpenAICompat | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/billing/keys/redeem": {
      "post": {
        "tags": [
          "billing"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyRedeemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyRedeemResponse"
                }
              }
            }
          },
          "402": {
            "description": "Payment Required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyRedeemResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyRedeemResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyRedeemResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/billing/keys/revoke": {
      "post": {
        "tags": [
          "billing"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyRevokeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/billing/operator/quotes/{quoteId}/approve": {
      "post": {
        "tags": [
          "billing"
        ],
        "parameters": [
          {
            "name": "quoteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorApproveResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/billing/operator/keys": {
      "post": {
        "tags": [
          "billing"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyIssueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyIssueResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiKeyIssueRequest": {
        "required": [
          "tenant",
          "label"
        ],
        "type": "object",
        "properties": {
          "tenant": {
            "type": [
              "null",
              "string"
            ]
          },
          "label": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ApiKeyIssueResponse": {
        "required": [
          "api_key",
          "key_prefix",
          "tenant",
          "label",
          "note"
        ],
        "type": "object",
        "properties": {
          "api_key": {
            "type": "string"
          },
          "key_prefix": {
            "type": "string"
          },
          "tenant": {
            "type": "string"
          },
          "label": {
            "type": [
              "null",
              "string"
            ]
          },
          "note": {
            "type": "string"
          }
        }
      },
...
      "ApiKeyRedeemRequest": {
        "required": [
          "session_id",
          "label"
        ],
        "type": "object",
        "properties": {
          "session_id": {
            "type": [
              "null",
              "string"
            ]
          },
          "label": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ApiKeyRedeemResponse": {
        "required": [
          "redeemed",
          "status",
          "tenant",
          "plan_id",
          "api_key",
          "key_prefix",
          "note"
        ],
        "type": "object",
        "properties": {
          "redeemed": {
            "type": "boolean"
          },
          "status": {
            "type": "string"
          },
          "tenant": {
            "type": [
              "null",
              "string"
            ]
          },
          "plan_id": {
            "type": [
              "null",
              "string"
            ]
          },
          "api_key": {
            "type": [
              "null",
              "string"
            ]
          },
          "key_prefix": {
            "type": [
              "null",
              "string"
            ]
          },
          "note": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ApiKeyRevokeRequest": {
        "required": [
          "tenant",
          "key_prefix"
        ],
        "type": "object",
        "properties": {
          "tenant": {
            "type": [
              "null",
              "string"
            ]
          },
          "key_prefix": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "OperatorApproveResponse": {
        "required": [
          "approved",
          "quote_id",
          "status"
        ],
        "type": "object",
        "properties": {
          "approved": {
            "type": "boolean"
          },
          "quote_id": {
            "type": "string"
          },
          "status": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}